Make it more difficult to overwrite existing packages with aliases.
Packages were being built and then promptly overwritten with an
alias package pointing at themselves in some cases.  This turns
those cases into errors earlier rather than later.

+++ pkg.alias.sh
@@ -8,6 +8,12 @@ add_flag --required alias "Name of the virtual package to create."
 add_flag --required target "Name of actual package to be installed"
 parse_flags
 
+if [[ -e "$_REPO_LOCAL_PATH/$F_alias.done" ]]
+then
+  echo "$(basename "$0"): package '$F_alias' already exists, aborting" >&2
+  exit 1
+fi
+
 tar -c -T /dev/null -z -f "$_REPO_LOCAL_PATH/$F_alias.tar.gz"
 echo "alias" > "$_REPO_LOCAL_PATH/$F_alias.version"
 echo "$F_target" > "$_REPO_LOCAL_PATH/$F_alias.dependencies"
